feat: Add SVG icons for Apple and Google
This commit adds SVG icons for Apple and Google in the `app/static/images` directory. The `apple-black.svg` and `google.svg` files contain the SVG code for the respective icons. These changes enhance the project by providing visually appealing icons for Apple and Google, which can be used in various parts of the application.